// packages/editor-core/src/model-types-primitives.d.ts
export interface LUTLayer {
    path: string;
    intensity: number;
}